Add skin support for minibosses
